+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-streams",
-  "version": "13.2.2",
+  "version": "13.3.0",
   "description": "Simple Streams for React",
   "main": "dist/react-streams.js",
   "module": "dist/react-streams.esm.js",
