+++ src/operators/spreadMap.ts
@@ -1,0 +1,4 @@
+import { map } from "rxjs/operators"
+
+export const spreadMap = (overrides = {}) =>
+  map(value => ({ ...value, ...overrides }))
