+++ src/utils/curry.ts
@@ -1,0 +1,2 @@
+export const curry = fn => (...args) =>
+  args.length < fn.length ? curry(fn.bind(null, ...args)) : fn(...args)
