+++ src/utils/isNotPlan.ts
@@ -1,0 +1,3 @@
+import { isObservable } from "rxjs"
+
+export const isNotPlan = x => isObservable(x) && !(x instanceof Function)
